[DEV] Modify Qt's FlowLayout to handle vertical orientation
By default the FlowLayout will fill the available horizontal space before
going to the next line.

It was not possible to fill the available vertical space first and then
go the next column.

While this patch is not pretty and probably not optimized, it is now possible
to do so.

How to use it:

- When declaring a FlowLayout, specify its orientation :
   layout = FlowLayout(wid, orientation=Qt.Vertical)
- Connect the widthChanged signal to the setMinimumWidth function of the container:
    layout.widthChanged.connect(wid.setMinimumWidth)